+++ kiito_surveytool/src/components/Content.tsx
@@ -1,7 +1,7 @@
 import {   Dispatch, SetStateAction, useState } from 'react'
 import {  Stack } from '@mui/material'
-import { Qcategories, Kyselysivu } from '.'
-import {Tulokset} from '../tulokset/Tulokset'
+import { FrontButtons, Querypage } from '.'
+import {Result} from './Result'
 
 
 let retAns :number[][][]
@@ -28,10 +28,10 @@ const Content = (props:any) => {
       color: 'black',
       height: { xs: '100vh', sm: '100vh', md: '150vh'}
     }}>
-    {valittu === -1 && <Qcategories setValittu={ (data: SetStateAction<number>) => setValittu(data) }
+    {valittu === -1 && <FrontButtons setValittu={ (data: SetStateAction<number>) => setValittu(data) }
      data = {data} setName={setName} name = {name} localData={props.localData}/>}
-    {valittu > -1 && <Kyselysivu sivu = {valittu} setValittu = {setValittu} data = {data} />}
-    {valittu < -1 && <Tulokset sivu={(valittu*-1) - 2} setValittu = {setValittu} data = {data} localData={props.localData} answers={ans}/>}
+    {valittu > -1 && <Querypage sivu = {valittu} setValittu = {setValittu} data = {data} />}
+    {valittu < -1 && <Result sivu={(valittu*-1) - 2} setValittu = {setValittu} data = {data} localData={props.localData} answers={ans}/>}
 
     </Stack>
