--- kiito_surveytool/src/components/Kyselysivu.tsx
@@ -1,14 +1,0 @@
-import { Stack } from '@mui/material'
-import KyselyContent from './KyselyContent'
-
-const Kyselysivu = (props : any) => {
-  const { data } = props;
-
-  return (
-    
-      <KyselyContent sivu = {props.sivu} setValittu = {props.setValittu} data = {data}/>
-    
-  )
-}
-
-export default Kyselysivu
