+++ kiito_surveytool/src/components/ResultGraph.css
