+++ kiito_surveytool/src/index.css
@@ -12,7 +12,7 @@
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image: url("src/assets/Tausta.jpg");
+  background-image: url("/public/Tausta.jpg");
   background-repeat: no-repeat;
   background-size: cover;
   background-position: top center;
