--- kiito_surveytool/src/utils/constants.tsx
@@ -1,7 +1,0 @@
-export const categories = [
-    { name: 'Kiertävät raaka-aineet',  },
-    { name: 'Tuote palveluna'  },
-    { name: 'Jakamisalustat'  },
-    { name: 'Elinkaaren pidentäminen' },
-    { name: 'Resurssien talteenotto' },
-  ];
